% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segregatr-package.R
\docType{package}
\name{segregatr}
\alias{segregatr}
\title{segregatr: Segregation Analysis for Identifying Pathogenic Variants}
\description{
Quantifying the evidence of co-segregation in the evaluation of putative
pathogenic variants
}
