% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegregation.R
\name{plotSegregation}
\alias{plotSegregation}
\title{Pedigree plot for segregation analysis}
\usage{
plotSegregation(
  x,
  affected = NULL,
  unknown = NULL,
  proband = NULL,
  carriers = NULL,
  homozygous = NULL,
  noncarriers = NULL,
  cex = 1,
  margins = 1,
  pos.geno = "bottom",
  pos.arrow = "bottomleft",
  ...
)
}
\arguments{
\item{x}{A \code{\link[pedtools:ped]{pedtools::ped()}} object.}

\item{affected}{The affected pedigree members.}

\item{unknown}{Pedigree members with unknown affection status.}

\item{proband}{The ID label of the proband. This person must also be in both
\code{carriers} and \code{affected}.}

\item{carriers}{A character vector (or coercible to such), containing the ID
labels of pedigree members known to carry one copy of the variant in
question.}

\item{homozygous}{A character vector (or coercible to such), containing the
ID labels of pedigree members known to carry two copies of the variant in
question.}

\item{noncarriers}{A character vector (or coercible to such), containing the
ID labels of pedigree members known \emph{not} to carry the variant in question.}

\item{cex, margins}{Arguments passed on to \code{\link[pedtools:plot.ped]{pedtools::plot.ped()}}.}

\item{pos.geno}{Position of genotype labels relative to pedigree symbols;
either "bottom" (default), "topleft" or "topright".}

\item{pos.arrow}{Position of the proband arrow; either "bottomleft",
"bottomright", "topleft" or "topright".}

\item{...}{Optional plot parameters passed on to \code{\link[pedtools:plot.ped]{pedtools::plot.ped()}}.}
}
\description{
Plots a pedigree showing the segregation of a variant.
}
\examples{

x = cousinPed(1)
plotSegregation(x, affected = c(3,6,8), unknown = 1, carrier = 2:5,
                homozygous = 8, noncarriers = 6:7, proband = 8)

# Different placement of genotypes and proband arrow
plotSegregation(x, affected = c(3,6,8), unknown = 1, carrier = 2:5,
                homozygous = 8, noncarriers = 6:7, proband = 8,
                pos.geno = "topleft", pos.arrow = "bottomright")

}
